+++ app/Models/MenuItem.php
@@ -74,6 +74,14 @@ class MenuItem extends Model
     }
 
     /**
+     * @return HasMany<OrderItem, $this>
+     */
+    public function orderItems(): HasMany
+    {
+        return $this->hasMany(OrderItem::class);
+    }
+
+    /**
      * @return BelongsToMany<ModifierGroup, $this>
      */
     public function modifierGroups(): BelongsToMany