+++ app/Models/ServicePoint.php
@@ -80,6 +80,14 @@ class ServicePoint extends Model
     }
 
     /**
+     * @return HasMany<Order, $this>
+     */
+    public function orders(): HasMany
+    {
+        return $this->hasMany(Order::class);
+    }
+
+    /**
      * @return HasOne<TableSession, $this>
      */
     public function activeTableSession(): HasOne